Condensed all contollers to events folder
