+++ bot/client.js
@@ -9,56 +9,36 @@ if (process.env.NODE_ENV === 'production' && !process.env.DEBUG_MODE) process.en
 const client = new Discord.Client();
 
 // Pre-load controllers
-const controllers = require('./controllers')();
 const events = require('./events')();
 
 // Sorts and runs events
-function eventPicker(eventObj) {
-  // Check if the eventObj is 'message', and for ! prefix on message to ensure it is a command
-  if (eventObj.constructor.name.toLowerCase() === 'message' && eventObj.content.charAt(0) === '!') {
-    util.log('Command message received', eventObj.content, 0);
-
-    // Build basic help string
-    let helpString = 'v1.4.0 Discovered Commands:\n\n\t**<> - Required Item\t\t[] - Optional Item**';
-
+function eventPicker(eventName, eventObj) {
+  // Check if the event is a message event, and for ! prefix on message to ensure it is a command
+  if (eventName === 'message' && eventObj.content.charAt(0) === '!') {
+    util.log(`${eventName} event received`, 0);
     // Process message against every controller
-    Object.keys(controllers).forEach((key) => {
+    Object.keys(events).forEach((key) => {
       // Instantiate the controller
-      const controllerInstance = new controllers[key](eventObj);
-      util.log('Controller instance', controllerInstance, 5);
+      const eventInstance = new events[key](eventObj);
+      util.log('Controller instance', eventInstance, 5);
       // Runs commands after constructor is called
-      controllerInstance.run();
-
-      // Loop through commands if help command and add to string
-      if (eventObj.content.toLowerCase() === '!help') {
-        Object.keys(controllerInstance.commands).forEach((commandKey) => {
-          const commandInstance = controllerInstance.commands[commandKey];
-          // Check if command should be shown in help menu
-          if (commandInstance.showWithHelp) {
-            if (commandInstance.adminOnly && isAdmin(eventObj.member)) {
-              helpString += `\n\n \`${commandInstance.example}\` **- Admin Only** \n\t ${commandInstance.description}`;
-            } else if (commandInstance.adminOnly && !isAdmin(eventObj.member)) {
-              helpString += '';
-            } else {
-              helpString += `\n\n \`${commandInstance.example}\` \n\t ${commandInstance.description}`;
-            }
-          }
-        });
+      if (eventInstance.eventName.toLowerCase() === eventName.toLowerCase()) {
+        console.log('Fired');
+        eventInstance.run();
       }
     });
-
-    // If help command called, display string
-    if (eventObj.content.toLowerCase() === '!help') {
-      eventObj.reply(helpString);
-    }
+  // If not a message event, treat as a normal event
   } else {
+    util.log(`${eventName} event received`, 0);
     // Process event against every controller
     Object.keys(events).forEach((key) => {
       // Instantiate the controller
       const eventInstance = new events[key](eventObj);
       util.log('Event instance', eventInstance, 5);
       // Runs commands after constructor is called
-      eventInstance.run();
+      if (eventInstance.eventName.toLowerCase() === eventName.toLowerCase()) {
+        eventInstance.run();
+      }
     });
   }
 }
@@ -70,12 +50,17 @@ client.on('ready', () => {
 
 // Create an event listener for new guild members
 client.on('guildMemberAdd', (member) => {
-  eventPicker(member);
+  eventPicker('guildMemberAdd', member);
+});
+
+// Fire whenever a user's presence changes
+client.on('presence', (user) => {
+  eventPicker('presence', user);
 });
 
 // Listen for messages
 client.on('message', (message) => {
-  eventPicker(message);
+  eventPicker('message', message);
 });
 
 client.login(process.env.TOKEN);
